refactor: fix import and writing solution
general calling writing solution
