+++ python/constants/__init__.py
@@ -6,4 +6,5 @@ from .plan_query import PLAN_QUERY, PLAN_PROGRESS_QUERY
 from .submit_query import (RECENT_SUBMISSIONS_QUERY, RECENT_AC_SUBMISSIONS_QUERY, USER_PROFILE_QUESTIONS_QUERY,
                            PROGRESS_SUBMISSIONS_QUERY, MY_SUBMISSION_DETAIL_QUERY)
 from .user_query import USER_PROFILE_PUBLIC_QUERY
-from .code_templates import TESTCASE_TEMPLATE_PYTHON, TESTCASE_TEMPLATE_PYTHON_TESTCASES, SOLUTION_TEMPLATE_PYTHON
+from .code_templates import (TESTCASE_TEMPLATE_PYTHON, TESTCASE_TEMPLATE_PYTHON_TESTCASES,
+                             SOLUTION_TEMPLATE_PYTHON, SOLUTION_TEMPLATE_GOLANG,)
