+++ python/lc_libs/__init__.py
@@ -3,5 +3,6 @@ from .question import get_question_info, get_question_desc, get_question_code, g
     extract_outputs_from_md, get_questions_by_key_word, CATEGORY_SLUG, LANGUAGE_SLUG
 from .submission import check_accepted_submission, get_submission_detail, check_accepted_submission_all
 from .user import check_user_exist
-from .writers import write_problem_md, write_testcase, write_solution_python, write_solution_golang
+from .python_writer import write_problem_md, write_testcase, write_solution_python
+from .golang_writer import write_solution_golang
 from .study_plan import get_user_study_plans, get_user_study_plan_progress
