+++ python/lc_libs/python_writer.py
@@ -4,7 +4,8 @@ import time
 import traceback
 from collections import defaultdict
 from importlib.util import spec_from_file_location, module_from_spec
-from constants import TESTCASE_TEMPLATE_PYTHON, TESTCASE_TEMPLATE_PYTHON_TESTCASES, SOLUTION_TEMPLATE_PYTHON
+
+from python.constants import TESTCASE_TEMPLATE_PYTHON, TESTCASE_TEMPLATE_PYTHON_TESTCASES, SOLUTION_TEMPLATE_PYTHON
 
 
 def write_problem_md(question_id: str, question_name: str, desc: str) -> str:
@@ -220,7 +221,7 @@ def __extract_process_input_from_method(cs_map, modify_in_place, import_libs, me
                 inputs += f"head{idx}"
                 idx += 1
         elif "Node" in str(v.annotation) and "Node" in cs_map and "neighbors" in cs_map["Node"][0][1]:
-            # speical handle Neighbour Nodes
+            # special handle Neighbour Nodes
             exists = True
             add_lib = "from python.object_libs import list_relation_to_node_neigh"
             if "List[" in str(v.annotation):
@@ -257,7 +258,7 @@ def __extract_process_input_from_method(cs_map, modify_in_place, import_libs, me
             remain += ("        res = self.{}({})\n        return linked_list_to_list(res)"
                        .format(func_name, inputs))
     elif "Node" in str(return_anno) and "Node" in cs_map and "neighbors" in cs_map["Node"][0][1]:
-        # speical handle Neighbour Nodes
+        # special handle Neighbour Nodes
         add_lib += ", node_neigh_to_list_relation" if exists else \
             "from python.object_libs import node_neigh_to_list_relation"
         if "List[" in str(return_anno):
@@ -451,288 +452,29 @@ def __write_solution_python_backup(code: str):
     )
 
 
-def write_solution_python(code: str, default: bool = True) -> str:
-    if not default:
-        # submission code, not template code
-        if "class Solution" in code:
-            return "\n".join(code.split("class Solution")[-1].split("\n")[1:])
-        return code
+def write_solution_python(code_template: str, code: str = None) -> str:
     try:
-        cs_map, defined_class, rest = __process_code(code)
-        modify_in_place = "Do not return anything" in code
+        cs_map, defined_class, rest = __process_code(code_template)
+        modify_in_place = "Do not return anything" in code_template
         if len(cs_map) == 1:
             import_libs, process_input = __finalize_solution_code_with_single_class(cs_map, modify_in_place)
         else:
             import_libs, process_input = __finalize_solution_code_complex(cs_map, modify_in_place)
+        if code:
+            # submission code, not template code
+            if "class Solution" in code:
+                last_part = "\n".join(code.split("class Solution")[-1].split("\n")[1:])
+            else:
+                last_part = code
+        else:
+            last_part = ("" if not rest or "(self" in rest[0] else "\n") + "\n".join(rest)
         return SOLUTION_TEMPLATE_PYTHON.format(
             "".join(import_libs) + ("\n" if import_libs and defined_class else "") +
             (("\n" if defined_class else "") + "\n".join(defined_class) + ("\n" if defined_class else "")),
             process_input,
-            ("" if not rest or "(self" in rest[0] else "\n") + "\n".join(rest)
+            last_part
         )
     except Exception as e:
         print("Exception raised:", e)
         traceback.print_exc()
-    return __write_solution_python_backup(code)
-
-
-def write_solution_golang(code_default: str, problem_id: str, default: bool = True, code: str = "") -> str:
-    base_str = ("package problem{}\n\n"
-                "import (\n"
-                "{}\n"
-                ")\n\n"
-                "{}\n\n"
-                "func Solve(input string) {}\n"
-                "\tvalues := strings.Split(input, \"\\n\")\n"
-                "{}\n{}\n"
-                "\treturn {}({})\n{}\n"
-                )
-
-    def process_inputs(input_str: str, struct_dict: dict, struct_func: bool = False) -> (set, str, str, str):
-        res = []
-        imports_libs = set()
-        json_parse = []
-        variables = []
-        if input_str.strip() == "":
-            return set(), "", "", ""
-        splits = input_str.split(",")
-        first = True
-        list_type_vars = []
-        for i, s in enumerate(splits):
-            ss = s.split(" ")
-            tmp_ss = []
-            for tmp_s in ss:
-                if tmp_s.strip() != "":
-                    tmp_ss.append(tmp_s)
-            variables.append(tmp_ss[0])
-            if first:
-                list_type_vars.append([])
-                res.append("\tvar ")
-            list_type_vars[-1].append(tmp_ss[0])
-            if len(tmp_ss) != 2:
-                res.append(tmp_ss[0])
-                res.append(", ")
-                first = False
-            else:
-                list_type_vars[-1].append(tmp_ss[1])
-                res.append(tmp_ss[0])
-                res.append(" ")
-                tp = tmp_ss[1]
-                res.append(tp)
-                res.append("\n")
-                first = True
-        cnts = 0
-        if struct_func:
-            variables = []
-        for i, vars_type in enumerate(list_type_vars):
-            vrs, tp = vars_type[:-1], vars_type[-1]
-            if (tp.startswith("*") and tp[1:] in struct_dict) or tp in struct_dict:
-                for var in vrs:
-                    print(var)
-            elif struct_func:
-                imports_libs.add("\t\"encoding/json\"")
-                imports_libs.add("\t\"log\"")
-                for _ in vrs:
-                    variables.append(f"values[{cnts}].({tp})") if tp != "int" else variables.append(
-                        f"int(values[{cnts}].(float64))")
-                    cnts += 1
-            else:
-                match tp:
-                    case "*ListNode":
-                        for var in vrs:
-                            json_parse.append(f"\tvar {var}IntArray []int\n")
-                            json_parse.append(f"\tif err := json.Unmarshal([]byte(values[{i}]), &" + var +
-                                              "IntArray); err != nil {\n\t\tlog.Fatal(err)\n\t}\n")
-                            json_parse.append(f"\t{var} = IntArrayToLinkedList({var}IntArray)\n")
-                        imports_libs.add("\t. \"leetCode/golang/models\"")
-                        imports_libs.add("\t\"encoding/json\"")
-                        imports_libs.add("\t\"log\"")
-                    case "*TreeNode":
-                        for var in vrs:
-                            json_parse.append(f"\t{var} = ArrayToTree(values[{i}])\n")
-                        imports_libs.add("\t. \"leetCode/golang/models\"")
-                    case "*Node":
-                        if ("Left *Node" in code_default
-                                and "Right *Node" in code_default
-                                and "Next *Node" in code_default):
-                            for var in vrs:
-                                json_parse.append(f"\t{var} = ArrayToTree(values[{i}])\n")
-                            imports_libs.add("\t. \"leetCode/golang/tree_next\"")
-                        elif "Neighbors []*Node" in code_default:
-                            for var in vrs:
-                                json_parse.append("\tvar arr" + f"{i}" + " [][]int\n")
-                                json_parse.append(f"\tif err := json.Unmarshal([]byte(values[{i}]), &" + f"arr{i}" +
-                                                  "); err != nil {\n\t\tlog.Fatal(err)\n\t}\n")
-                                json_parse.append(f"\t{var} = ArrayRelationToNodeNeighbour(arr{i})\n")
-                            imports_libs.add("\t. \"leetCode/golang/node_neighbours\"")
-                            imports_libs.add("\t\"encoding/json\"")
-                            imports_libs.add("\t\"log\"")
-                        elif "/**\n" \
-                             " * Definition for a Node.\n" \
-                             " * type Node struct {\n" \
-                             " *     Val int\n" \
-                             " *     Next *Node\n" \
-                             " *     Random *Node\n" \
-                             " * }\n" \
-                             " */" in code_default:
-                            for var in vrs:
-                                json_parse.append("\tvar arr" + f"{i}" + " [][]interface{}\n")
-                                json_parse.append(f"\tif err := json.Unmarshal([]byte(values[{i}]), &" + f"arr{i}" +
-                                                  "); err != nil {\n\t\tlog.Fatal(err)\n\t}\n")
-                                json_parse.append(f"\t{var} = IntRandomArrayToNodeArray(arr{i})\n")
-                            imports_libs.add("\t. \"leetCode/golang/node_random\"")
-                            imports_libs.add("\t\"encoding/json\"")
-                            imports_libs.add("\t\"log\"")
-                    case _:
-                        for var in vrs:
-                            json_parse.append(f"\tif err := json.Unmarshal([]byte(values[{i}]), &" + var +
-                                              "); err != nil {\n\t\tlog.Fatal(err)\n\t}\n")
-                        imports_libs.add("\t\"encoding/json\"")
-                        imports_libs.add("\t\"log\"")
-        imports_libs.add("\t\"strings\"")
-        return imports_libs, "".join(res), "".join(json_parse), ", ".join(variables)
-
-    its = []
-    rts = []
-    func_names = []
-    structs_map = dict()
-    for i, line in enumerate(code_default.split("\n")):
-        line = line.strip()
-        if line.startswith("func "):
-            rts.append(line.split("{")[0].split(")")[-1].strip())
-            its.append(process_inputs(line.split("(")[1].split(")")[0], structs_map))
-            func_names.append(line.split("(")[0].split("func ")[-1].strip())
-        elif line.startswith("type ") and line.endswith(" struct {"):
-            struct_name = line[len("type "):-len(" struct {")]
-            structs_map[struct_name] = dict()
-            for tmp in code_default.split("\n"):
-                tmp = tmp.strip()
-                if tmp.startswith("func ") and (
-                        tmp.endswith(f") {struct_name} " + "{") or
-                        tmp.endswith(f") *{struct_name} " + "{")):
-                    tp0, tp1, tp2, tp3 = process_inputs(tmp.split("(")[1].split(")")[0],
-                                                        structs_map, True)
-                    rt = tmp.split("{")[0].split(")")[-1].strip()
-                    structs_map[struct_name]["construct"] = (tmp.split("(")[0].split("func ")[-1].strip(),
-                                                             (tp0, tp1, tp2, tp3.replace("values", "vals[0]")),
-                                                             rt)
-                elif tmp.startswith("func (") and struct_name in tmp.split(")")[0]:
-                    if "funcs" not in structs_map[struct_name]:
-                        structs_map[struct_name]["funcs"] = []
-                    tp0, tp1, tp2, tp3 = process_inputs(tmp.split("(")[2].split(")")[0],
-                                                        structs_map, True)
-                    rt = tmp.split("{")[0].split(")")[-1].strip()
-                    structs_map[struct_name]["funcs"].append((tmp.split("(")[1].split(")")[-1].strip(),
-                                                              (tp0, tp1, tp2, tp3.replace("values", "vals[i]")),
-                                                              rt))
-
-            import_set = set()
-            func_loop = ""
-            constructor = None
-            for d in structs_map.values():
-                if "funcs" in d:
-                    for name, its, rt in d["funcs"]:
-                        import_set.update(its[0])
-                        func_loop += ("\t\tcase \"{}\", \"{}\":\n"
-                                      "\t\t\t{}obj.{}({})\n").format(name[0].lower() + name[1:], name,
-                                                                     "res = nil\n\t\t\t" if rt == "" else "res = ",
-                                                                     name, its[3])
-                if "construct" in d:
-                    constructor = d["construct"]
-            build_body = ("\tvar opts []string\n" +
-                          "\tvar vals [][]interface{}\n" +
-                          "\tvar ans []interface{}\n" +
-                          "\tif err := json.Unmarshal([]byte(values[0]), &opts); err != nil {\n" +
-                          "\t\tlog.Println(err)\n" +
-                          "\t\treturn nil\n" +
-                          "\t}\n" +
-                          "\tif err := json.Unmarshal([]byte(values[1]), &vals); err != nil {\n" +
-                          "\t\tlog.Println(err)\n" +
-                          "\t\treturn nil\n" +
-                          "\t}\n" +
-                          "{}".format(
-                              "\tobj :=" + constructor[
-                                  0] + f"({constructor[1][3]})\n" if constructor is not None else "",
-                              ""
-                          ) +
-                          "\tans = append(ans, nil)\n" +
-                          "\tfor i := 1; i < len(opts); i++ {\n" +
-                          "\t\tvar res interface{}\n" +
-                          "{}".format(
-                              "\t\tswitch opts[i] {\n" +
-                              func_loop +
-                              "\t\tdefault:\n"
-                              "\t\t\tres = nil\n"
-                              "\t\t}\n"
-                          ) +
-                          "\t\tans = append(ans, res)\n"
-                          "\t}\n"
-                          )
-
-            return base_str.format(
-                problem_id,
-                "\n".join(sorted(import_set, key=lambda x: "\t" + x.split(" ")[-1] if x.startswith('\t. ') else x)),
-                code_default if default else code,
-                "interface{} {",
-                build_body,
-                "",
-                "ans",
-                "",
-                "}",
-            ).replace("return ans()", "return ans")
-    import_set = set()
-    for it in its:
-        import_set.update(it[0])
-
-    if len(rts) != 1 or rts[0] == "*TreeNode" or rts[0] == "*ListNode" or rts[0] == "*Node":
-        return_func_var = "{}({})".format(func_names[0],
-                                          ", ".join(list(zip(*its))[3]))
-        match rts[0]:
-            case "*TreeNode":
-                return_func_name = "TreeToArray"
-            case "*ListNode":
-                return_func_name = return_func_var + ".LinkedListToIntArray"
-                return_func_var = ""
-            case "*Node":
-                return_func_name = "ToBeImplemented"
-                if ("Left *Node" in code_default
-                        and "Right *Node" in code_default
-                        and "Next *Node" in code_default):
-                    return_func_name = "TreeNextToArray"
-                elif "Neighbors []*Node" in code_default:
-                    return_func_name = "NodeNeighbourToArrayRelation"
-                elif "/**\n" \
-                     " * Definition for a Node.\n" \
-                     " * type Node struct {\n" \
-                     " *     Val int\n" \
-                     " *     Next *Node\n" \
-                     " *     Random *Node\n" \
-                     " * }\n" \
-                     " */" in code_default:
-                    return_func_name = "NodeArrayToIntRandomArray"
-            case _:
-                return_func_name = ""
-
-        return base_str.format(
-            problem_id,
-            # string with . starts before other, othercase sort normal
-            "\n".join(sorted(import_set, key=lambda x: "\t" + x.split(" ")[-1] if x.startswith('\t. ') else x)),
-            code_default if default else code,
-            "interface{} {",
-            "\n".join(list(zip(*its))[1]),
-            "\n".join(list(zip(*its))[2]),
-            return_func_name,
-            return_func_var,
-            "}",
-        )
-    return base_str.format(
-        problem_id,
-        "\n".join(sorted(import_set, key=lambda x: "\t" + x.split(" ")[-1] if x.startswith('\t. ') else x)),
-        code_default if default else code,
-        "interface{} {",
-        "\n".join(list(zip(*its))[1]),
-        "\n".join(list(zip(*its))[2]),
-        func_names[0],
-        ", ".join(list(zip(*its))[3]),
-        "}",
-    )
+    return __write_solution_python_backup(code_template)
