+++ python/lc_libs/question.py
@@ -1,13 +1,14 @@
 import json
 import traceback
-import requests
-import markdown
-import html2text
 from typing import Optional, Mapping
 
-from constants import (LEET_CODE_BACKEND, QUESTION_INFO_QUERY, QUESTION_DESC_QUERY, QUESTION_CODE_QUERY,
-                       QUESTION_TESTCASE_QUERY, QUESTION_KEYWORDS_QUERY)
-from utils import general_request
+import html2text
+import markdown
+import requests
+
+from python.constants import (LEET_CODE_BACKEND, QUESTION_INFO_QUERY, QUESTION_DESC_QUERY, QUESTION_CODE_QUERY,
+                              QUESTION_TESTCASE_QUERY, QUESTION_KEYWORDS_QUERY)
+from python.utils import general_request
 
 CATEGORY_SLUG = {"all-code-essentials", "algorithms", "database"}
 LANGUAGE_SLUG = {"python3", "mysql"}
